% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSampleRNAdynamics.R
\name{eval.rates.vec}
\alias{eval.rates.vec}
\title{}
\usage{
eval.rates.vec(pre.frac, lab.frac, tp = 1, cl = NULL)
}
\arguments{
\item{pre.frac}{unlabled intron to exon ratios for multiple transcripts}

\item{lab.frac}{labled intron to exon ratios for the same transcripts}

\item{tp}{labeling time}

\item{cl}{cluster for parallelization (NULL if no parallelization)}
}
\value{
a matrix where each row corresponds to a transcripts. The first column corresponds to the degradation rate, and the second column corresponds to the processing rate. A third and fourth column containing degradation and processing rates for a second mimima are provided if necessary.
}
\description{
.. vectorized version of function optimization
}
\details{
.. content for \details{} ..
}
\author{
Micha Hersch
}
