% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSampleRNAdynamics.R
\name{moptim}
\alias{moptim}
\title{Performs function minimization on a bounded domain using a gradient method.}
\usage{
moptim(lpre.frac, llab.frac, init.par, lbm, ubm)
}
\arguments{
\item{lpre.frac}{log unlabeled intron to exon ratio}

\item{llab.frac}{log labeled intron to exon ratio}

\item{init.par}{initial parameter value for searcu}

\item{lbm}{lower bound}

\item{ubm}{upper bound}
}
\value{
Locations of two mimimas
}
\description{
Looks for two minimas and returns them both.
}
\author{
Micha Hersch
}
