% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSampleRNAdynamics.R
\name{singleSampleRateEstimation}
\alias{singleSampleRateEstimation}
\title{}
\usage{
singleSampleRateEstimation(TPM.data)
}
\arguments{
\item{TPM.data}{a data frame containing the columns with the following
self-explanatory names:
\itemize{
\item unlabeled.intron
\item unlabeled.exon
\item labeled.intron
\item labeled.exon
}}
}
\value{
A TPM.data converted to a data.table with the following additional columns
\itemize{
\item prod.rate the synthesis (or production) rate, in log (base e)
\item proc.rate the processing rate, also in log
\item deg.rate the degradation rate, also in log
\item pre.frac the unlabeled (pre-existing) intro to exon ratio
\item lab.frac the labeled (pre-existing) intro to exon ratio
\item solvable "C" if the rates were estimated by finding the zero of
the function (eq 13 of paper), "A" if the rates were estimated by minimization,
"B" if the rates cannont be estimated.
\item twosols: a boolean indicating whether two solutions have been found
}
}
\description{
Estimates RNA synthesis, processing and degradation rates from 4sU metabolic
labeling experiments with a pre-existing and labeled RNA quantification of
exons and introns.
}
\details{
This function implements the method described in Hersch et al,
"Estimating RNA dynamics using one time point for one sample in single-pulse
metabolic labeling experiment", 2022.

Note that this method can only estimate rates of transripts for which the ratio
of intron to exon expression is below 1 and is higher in the labeled RNA
than in the unlabeled RNA
(see the paper for the details).
}
\author{
Micha Hersch
}
