% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSampleRNAdynamics.R
\name{solve.rates}
\alias{solve.rates}
\title{}
\usage{
\method{solve}{rates}(frac, t = 1)
}
\arguments{
\item{frac}{a vector of size 2 containing :
\enumerate{
\item the prexisting intron to exon ratio
\item the labeled intron to exon ratio
}}

\item{t}{the labeling time}
}
\value{
a vector containing:
\enumerate{
\item the log degradation rate
\item the log processing rate
the rates in units given by the inverse of the units of t
If their is only one solution the vector is of size 2. If there are two
solutions, the vector is of
size 4 (degr.rate1, proc.rate1, degr.rate2, proc.rate2).
If there are no solution the vector c(NA,NA) is returned.
}
}
\description{
Extracts processing and degradation rates from unlabled (pre-existing) and labeled intro to exon ratios for a single transcript.
}
\details{
See reference paper for the details of the method. Ratios expected to be smaller that one and labeled ratio is expected to be larger than the unlabeled ratio.
}
\author{
Micha Hersch
}
